Add functions related to constructing backups from diffs
Given a seed day and a set of diffs, calculate all the sequences of diffs
for which the first has a from day matching the seed and the subsequent
ones have from day matching the to day of the previous diff. From such a
set of sequences, we can work out which backup days can be created and
which diffs are necessary to do so.