Avoid some actions throwing exceptions
The checking of a remote against its hash may fail because the local file
containing the hash may have been deleted. This is better interpreted as
the hash failing to match.

When removing a remote, we also remove the local copy of the hash. We
don't care if it has already been deleted.