Address some review comments
Thank you to the directors of Concert Audio Technologies Ltd for the
suggestions.

Fix a nasty inefficiency in SFTP.hGetUptoMark. After being careful to
store the buffered characters in reverse order, so that new ones could be
efficiently added with cons, and tests for the mark could be made
efficiently using isPrefixOf, a daft use of length crept in to check
whether the buffer and the mark were of equal length. This has been
replaced by a test for a null list resulting by dropping the length of the
known no-longer list from the potentially longer one. Also move the
final reversing of the list to make it more obvious that it is
performed only once.

Remove a gratuitous use of the function applicative to hack argument
passing.

Distinguish the subsiduary functions of the BackupDir class from the core
ones by making them outer level definitions and adding type constraints.

Make explicit which functions are used in imports.