Avoid explicit passing of the SFTP context to SFTP commands
The withSFTP function created a context which it passed to the commands
it runs. This commit removes the SFTP type that wrapped the context,
instead introducing an SFTP monad which provides for the implicit passing
of the context.

Internall, the StateT monad transformer is used to pass the context. We
could have used ReaderT, but there are further plans to pass other,
changing information via the state.