Improve the exception handling within the SFTP module
Previously, fail was called to signal an error, raising an exception, which
could be caught only outside of the withSFTP call, and not within the inner
code that ran the commands. (In fact, with some fiddling, it was possible
to catch the exception, but the inner state change was lost for the
command that failed, and any further commands would receive output from
the shell command out of sync)

Now we use the ExceptT monad transformer to support exceptions that don't
lose the inner state. They can be caught with catchT and further commands
executed.

This also gives us the freedom to implement withSFTPE, a variant of
withSFTP, which returns the result of the last command wrapped with
Either, rather than throwing exceptions. withSFTP's behaviour is
unchanged.